Initial Seeding from Distribution from 2021-04-09
This was a custom version that I configured on 2021-04-09.  I had removed some more experimental code, packaged in lsdl2019 and font resources, and wrote a set of tutorials.  Otherwise, it substantially matches minicubes/ as it appears on my computer.  Longer term, I will need to include a configuration system so that I can work more easily with the git repository version, and my own system configuration.  I may also include some diagnostics for SDL2, so that the problems that Sam had (for instance) with (probably SDL2) can be inquired into.